+++ CHANGELOG.md
@@ -5,6 +5,11 @@ Version numbering represents the Swift version, plus a running number representi
 You can also refer to commit logs to get details on what was implemented, fixed and improved.
 
 
+### #.#
+
+- The convenience `request(forURL:)` method by default no longer uses locally cached data
+
+
 ### 2.2.5
 
 - Use ephemeral NSURLSession by default; fixes #96